5807 mapping properties to dataLayer
