+++ assets/templates/partials/census/id-datestamp.tmpl
@@ -1,12 +1,12 @@
 <dl class="ons-metadata metadata__list ons-grid ons-grid--gutterless ons-u-cf ons-u-mb-m" title="{{ localise "CensusDatasetLandingPageMetadataLabel" .Language 1 }}" aria-label="{{ localise "CensusDatasetLandingPageMetadataLabel" .Language 1 }}">
     <div class="ons-grid__col ons-col-12@m">
         <dt class="ons-metadata__term ons-u-mr-xs">{{ localise "DatasetID" .Language 1 }}</dt>
-        <dd class="ons-metadata__value ons-u-f-no">{{ .ID }}</dd>
+        <dd class="ons-metadata__value ons-u-f-no">{{ .DatasetId }}</dd>
     </div>
     {{ if .DatasetLandingPage.HasOtherVersions }}
         <div class="ons-grid__col ons-col-4@m ons-u-mt-xs">
             <dt class="ons-metadata__term ons-u-mr-xs">{{ localise "Released" .Language 1 }}:</dt>
-            <dd class="ons-metadata__value ons-u-f-no"><time datetime="{{ .InitialReleaseDate }}">{{ dateFormat .InitialReleaseDate }}</time></dd>
+            <dd class="ons-metadata__value ons-u-f-no"><time datetime="{{ .ReleaseDate }}">{{ dateFormat .ReleaseDate }}</time></dd>
         </div>
         <div class="ons-grid__col ons-col-8@m ons-u-mt-xs">
             <dt class="ons-metadata__term ons-u-mr-xs">{{ localise "LastUpdated" .Language 1 }}:</dt>
@@ -17,7 +17,7 @@
     {{ else }}
         <div class="ons-grid__col ons-col-12@m ons-u-mt-xs">
             <dt class="ons-metadata__term ons-u-mr-xs">{{ localise "Released" .Language 1 }}:</dt>
-            <dd class="ons-metadata__value ons-u-f-no"><time datetime="{{ .InitialReleaseDate }}">{{ dateFormat .InitialReleaseDate }}</time></dd>
+            <dd class="ons-metadata__value ons-u-f-no"><time datetime="{{ .ReleaseDate }}">{{ dateFormat .ReleaseDate }}</time></dd>
         </div>
     {{ end }}
 </dl>
